+++ tests/advanced/test_custom_heuristics.py
@@ -13,7 +13,7 @@ random.seed(123)    # 50 terminal
 import steinerpy.config as cfg                              # to ensure heuristic is set to "preprocess"
 from steinerpy.context import Context                       # helper to load graph and run algorithms
 from steinerpy.library.graphs.graph import GraphFactory
-from steinerpy.algorithms.common import Common
+from steinerpy.algorithms.common import CustomHeuristics
 
 # Visualize things
 # cfg.Animation.visualize = True
@@ -47,13 +47,17 @@ class TestGenerateHeuristics(unittest.TestCase):
         # print(n, goal)
         return 0
 
+    def setUp(self):
+        self.old_binding = CustomHeuristics.h_func
+        self.old_setting = cfg.Algorithm.sstar_heuristic_type
+        cfg.Algorithm.sstar_heuristic_type = "custom"
+        CustomHeuristics.bind(self.my_cust_h_func)
+
+    def tearDown(self):
+        cfg.Algorithm.sstar_heuristic_type = self.old_setting  
+        CustomHeuristics.bind(self.old_binding)
+
     def test_load_heuristics_from_disk_run_sstar(self):
-
-        # change heuristic setting
-        cfg.Algorithm.sstar_heuristic_type = "custom"    
-
-        # bind custom_heuristics function
-        Common.custom_heuristics = self.my_cust_h_func
 
         # create context handler and run S*-mm
         context = Context(graph, T)